encoder implementation
a bad man with middle name francis told me that we did have enough
breakouts. his henchman had told me we didnt so i did the other
implementation first. that had to be scrapped so now here we go
